# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "记账 API。部分业务错误使用 HTTP 200，通过 status/msg 判断结果。",
        "title": "Jiezhang API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/payees": {
            "get": {
                "security": [
                    {
                        "AppID": [],
                        "SessionKey": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payees"
                ],
                "summary": "收款方列表",
                "operationId": "PayeesHandler_List",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "账本 ID；未提供时使用用户默认账本",
                        "name": "account_book_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功；兼容接口也可能在 HTTP 200 的 status/msg 中返回业务错误",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/types.PayeeListItem"
                            }
                        }
                    },
                    "500": {
                        "description": "服务错误",
                        "schema": {
                            "$ref": "#/definitions/types.APIResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "AppID": [],
                        "SessionKey": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payees"
                ],
                "summary": "创建收款方，Body: `{name}` 或 `{payee: {name}}`",
                "operationId": "PayeesHandler_Create",
                "parameters": [
                    {
                        "description": "名称，可使用 name 或 payee.name",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.PayeeWriteRequest"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "账本 ID；未提供时使用用户默认账本",
                        "name": "account_book_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功；兼容接口也可能在 HTTP 200 的 status/msg 中返回业务错误",
                        "schema": {
                            "$ref": "#/definitions/model.Payee"
                        }
                    },
                    "400": {
                        "description": "请求失败",
                        "schema": {
                            "$ref": "#/definitions/types.APIResponse"
                        }
                    },
                    "500": {
                        "description": "服务错误",
                        "schema": {
                            "$ref": "#/definitions/types.APIResponse"
                        }
                    }
                }
            }
        },
        "/payees/{id}": {
            "put": {
                "security": [
                    {
                        "AppID": [],
                        "SessionKey": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payees"
                ],
                "summary": "更新收款方",
                "operationId": "PayeesHandler_Update",
                "parameters": [
                    {
                        "description": "名称，可使用 name 或 payee.name",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.PayeeWriteRequest"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "账本 ID；未提供时使用用户默认账本",
                        "name": "account_book_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功；兼容接口也可能在 HTTP 200 的 status/msg 中返回业务错误",
                        "schema": {
                            "$ref": "#/definitions/model.Payee"
                        }
                    },
                    "400": {
                        "description": "请求失败",
                        "schema": {
                            "$ref": "#/definitions/types.APIResponse"
                        }
                    },
                    "404": {
                        "description": "请求失败",
                        "schema": {
                            "$ref": "#/definitions/types.APIResponse"
                        }
                    },
                    "500": {
                        "description": "服务错误",
                        "schema": {
                            "$ref": "#/definitions/types.APIResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "AppID": [],
                        "SessionKey": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payees"
                ],
                "summary": "删除收款方",
                "operationId": "PayeesHandler_Delete",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "账本 ID；未提供时使用用户默认账本",
                        "name": "account_book_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功；兼容接口也可能在 HTTP 200 的 status/msg 中返回业务错误",
                        "schema": {
                            "$ref": "#/definitions/types.APIResponse"
                        }
                    },
                    "400": {
                        "description": "请求失败",
                        "schema": {
                            "$ref": "#/definitions/types.APIResponse"
                        }
                    },
                    "404": {
                        "description": "请求失败",
                        "schema": {
                            "$ref": "#/definitions/types.APIResponse"
                        }
                    },
                    "500": {
                        "description": "服务错误",
                        "schema": {
                            "$ref": "#/definitions/types.APIResponse"
                        }
                    }
                }
            }
        },
...
        "/settings/feedback": {
            "post": {
                "security": [
                    {
                        "AppID": [],
                        "SessionKey": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Settings"
                ],
                "summary": "提交反馈，Body: `{content, type}`",
                "operationId": "SettingsHandler_Feedback",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "账本 ID；未提供时使用用户默认账本",
                        "name": "account_book_id",
                        "in": "query"
                    },
                    {
                        "description": "请求体",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.FeedbackRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功；兼容接口也可能在 HTTP 200 的 status/msg 中返回业务错误",
                        "schema": {
                            "$ref": "#/definitions/types.APIResponse"
                        }
                    },
                    "500": {
                        "description": "服务错误",
                        "schema": {
                            "$ref": "#/definitions/types.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "model.Payee": {
            "type": "object",
            "properties": {
                "accountBookID": {
                    "type": "integer",
                    "format": "int64"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer",
                    "format": "int64"
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userID": {
                    "type": "integer",
                    "format": "int64"
                }
            }
        },
        "types.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "msg": {
                    "type": "string"
                },
                "session": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "types.FeedbackRequest": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "type": {
                    "type": "integer"
                }
            }
        },
...
        "types.PayeeListItem": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "types.PayeeWriteRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "payee": {
                    "type": "object",
                    "properties": {
                        "name": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "AppID": {
            "type": "apiKey",
            "name": "X-WX-APP-ID",
            "in": "header"
        },
        "SessionKey": {
            "type": "apiKey",
            "name": "X-WX-Skey",
            "in": "header"
        }
    }
}